+++ mirrord/layer/tests/self_connect.rs
@@ -19,13 +19,16 @@ async fn test_self_connect(dylib_path: &PathBuf) {
     let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap().to_string();
     println!("Listening for messages from the layer on {addr}");
-    let env = get_env(dylib_path.to_str().unwrap(), &addr);
+    let env = get_env_no_fs(dylib_path.to_str().unwrap(), &addr);
     let mut test_process =
         TestProcess::start_process(executable, application.get_args(), env).await;
 
     // Accept the connection from the layer and verify initial messages.
-    let mut layer_connection =
-        LayerConnection::get_initialized_connection(&listener, application.get_app_port()).await;
+    let mut layer_connection = LayerConnection::get_initialized_connection_with_port(
+        &listener,
+        application.get_app_port(),
+    )
+    .await;
     println!("Application subscribed to port, sending tcp messages.");
     assert!(layer_connection.is_ended().await);
     test_process.wait_assert_success().await;