+++ vscode-ext/src/extension.ts
@@ -1,17 +1,43 @@
 import { CoreV1Api, V1NamespaceList, V1PodList } from '@kubernetes/client-node';
-import { settings } from 'cluster';
+import { resolve } from 'path';
 import * as vscode from 'vscode';
-
+import YAML from 'yaml';
+const TOML = require('toml');
 const semver = require('semver');
 const https = require('https');
 const k8s = require('@kubernetes/client-node');
 const path = require('path');
 const os = require('os');
+const glob = require('glob');
 
 const LIBRARIES: { [platform: string]: [var_name: string, lib_name: string] } = {
 	'darwin': ['DYLD_INSERT_LIBRARIES', 'libmirrord_layer.dylib'],
 	'linux': ['LD_PRELOAD', 'libmirrord_layer.so']
 };
+
+const DEFAULT_CONFIG = `
+{
+    "accept_invalid_certificates": false,
+    "feature": {
+        "network": {
+            "incoming": "mirror",
+            "outgoing": true
+        },
+        "fs": "read",
+        "env": true
+    }
+}
+`;
+
+// Populate the path for the project's .mirrord folder
+const MIRRORD_DIR = function () {
+	let folders = vscode.workspace.workspaceFolders;
+	if (folders === undefined) {
+		throw new Error('No workspace folder found');
+	}
+	return vscode.Uri.joinPath(folders[0].uri, '.mirrord');
+}();
+
 const versionCheckEndpoint = 'https://version.mirrord.dev/get-latest-version';
 const versionCheckInterval = 1000 * 60 * 3;
 
@@ -24,90 +50,27 @@ function getK8sApi(): CoreV1Api {
 	return k8sConfig.makeApiClient(k8s.CoreV1Api);
 }
 
-async function changeSettings() {
-	let agentNamespace = globalContext.workspaceState.get<string>('agentNamespace', 'default');
-	let impersonatedPodNamespace = globalContext.workspaceState.get<string>('impersonatedPodNamespace', 'default');
-	let fileOps = globalContext.workspaceState.get<boolean>('fileOps', true);
-	let invalidCertificates = globalContext.workspaceState.get<boolean>('invalidCertificates', false);
-	let trafficStealing = globalContext.workspaceState.get<boolean>('trafficStealing', false);
-	let remoteDNS = globalContext.workspaceState.get<boolean>('remoteDNS', true);
-	let outgoingTraffic = globalContext.workspaceState.get<boolean>('outgoingTraffic', true);
-	let includeEnvironmentVariables = globalContext.workspaceState.get<string>('includeEnvironmentVariables', '*');
-	let excludeEnvironmentVariables = globalContext.workspaceState.get<string>('excludeEnvironmentVariables', '');
-	let telemetry = globalContext.workspaceState.get<boolean>('telemetry', true);
-
-	const options = ['Change namespace for mirrord agent (current: ' + agentNamespace + ')',
-	'Change namespace for impersonated pod (current: ' + impersonatedPodNamespace + ')',
-	'Toggle file operations (current: ' + (fileOps ? 'enabled' : 'disabled') + ')',
-	'Toggle invalid certificates (current: ' + (invalidCertificates ? 'enabled' : 'disabled') + ')',
-	'Toggle traffic stealing (current: ' + (trafficStealing ? 'enabled' : 'disabled') + ')',
-	'Toggle remote DNS (current: ' + (remoteDNS ? 'enabled' : 'disabled') + ')',
-	'Toggle outgoing traffic (current: ' + (outgoingTraffic ? 'enabled' : 'disabled') + ')',
-	'Include environment variables (current: ' + includeEnvironmentVariables + ')',
-	'Exclude environment variables (current: ' + excludeEnvironmentVariables + ')',
-	'Toggle telemetry (current: ' + telemetry + ')'];
-	vscode.window.showQuickPick(options).then(async setting => {
-		if (setting === undefined) {
-			return;
-		}
-		else if (setting.startsWith('Toggle telemetry')) {
-			globalContext.workspaceState.update('telemetry', !telemetry);
-		}
-		else if (setting.startsWith('Toggle file')) {
-			globalContext.workspaceState.update('fileOps', !fileOps);
-		}
-		else if (setting.startsWith('Toggle invalid certificates')) {
-			globalContext.workspaceState.update('invalidCertificates', !invalidCertificates);
-		}
-		else if (setting.startsWith('Toggle traffic stealing')) {
-			globalContext.workspaceState.update('trafficStealing', !trafficStealing);
-		}
-		else if (setting.startsWith('Toggle remote DNS')) {
-			globalContext.workspaceState.update('remoteDNS', !remoteDNS);
-		}
-		else if (setting.startsWith('Toggle outgoing traffic')) {
-			globalContext.workspaceState.update('outgoingTraffic', !outgoingTraffic);
-		}
-		else if (setting.startsWith('Include environment variables')) {
-			vscode.window.showInputBox({ prompt: 'Enter environment variables to include (separated by commas, * for all)', value: includeEnvironmentVariables }).then(async value => {
-				if (value === undefined) {
-					return;
-				}
-				globalContext.workspaceState.update('includeEnvironmentVariables', value);
-			});
-		}
-		else if (setting.startsWith('Exclude environment variables')) {
-			vscode.window.showInputBox({ prompt: 'Enter environment variables to exclude', value: excludeEnvironmentVariables }).then(async value => {
-				if (value === undefined) {
-					return;
-				}
-				globalContext.workspaceState.update('excludeEnvironmentVariables', value);
-			});
-		} else if (setting.startsWith('Change namespace')) {
-			const namespaces: {
-				response: any;
-				body: V1NamespaceList;
-			} = await getK8sApi().listNamespace();
-			const namespaceNames = namespaces.body.items.map(namespace => namespace.metadata!.name!);
-			vscode.window.showQuickPick(namespaceNames, { placeHolder: 'Select namespace' }).then(async namespaceName => {
-				if (namespaceName === undefined) {
-					return;
-				}
-				if (setting.startsWith('Change namespace for mirrord agent')) {
-					globalContext.workspaceState.update('agentNamespace', namespaceName);
-				} else if (setting.startsWith('Change namespace for impersonated pod')) {
-					globalContext.workspaceState.update('impersonatedPodNamespace', namespaceName);
-				}
-			});
-		}
-	});
+// Get the file path to the user's mirrord-config file, if it exists. 
+async function configFilePath() {
+	let fileUri = vscode.Uri.joinPath(MIRRORD_DIR, '?(*.)mirrord.+(toml|json|y?(a)ml)');
+	let files = glob.sync(fileUri.fsPath);
+	return files[0] || '';
 }
 
+async function openConfig() {
+	let path = await configFilePath();
+	if (!path) {
+		path = vscode.Uri.joinPath(MIRRORD_DIR, 'mirrord.json');
+		await vscode.workspace.fs.writeFile(path, Buffer.from(DEFAULT_CONFIG));
+	}
+	vscode.workspace.openTextDocument(path).then(doc => {
+		vscode.window.showTextDocument(doc);
+	});
+}
 
 async function toggle(context: vscode.ExtensionContext, button: vscode.StatusBarItem) {
 	let state = context.globalState;
 	if (state.get('enabled')) {
-		// vscode.debug.registerDebugConfigurationProvider('*', new ConfigurationProvider(), 2);
 		state.update('enabled', false);
 		button.text = 'Enable mirrord';
 	} else {
@@ -139,6 +102,31 @@ async function checkVersion(version: string) {
 	});
 }
 
+// Gets namespace from config file
+async function parseNamespace() {
+	let parsed = await parseConfigFile();
+	return parsed?.['target']?.['namespace'] || 'default';
+}
+
+async function parseConfigFile() {
+	let filePath: string = await configFilePath();
+	if (filePath) {
+		const file = (await vscode.workspace.fs.readFile(vscode.Uri.parse(filePath))).toString();
+		if (filePath.endsWith('json')) {
+			return JSON.parse(file);
+		} else if (filePath.endsWith('yaml') || filePath.endsWith('yml')) {
+			return YAML.parse(file);
+		} else if (filePath.endsWith('toml')) {
+			return TOML.parse(file);
+		}
+	}
+}
+
+async function isTargetInFile() {
+	let parsed = await parseConfigFile();
+	return (parsed && (typeof (parsed['target']) === 'string' || parsed['target']?.['path']));
+}
+
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -158,7 +146,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	buttons.toggle.command = toggleCommandId;
 
 	const settingsCommandId = 'mirrord.changeSettings';
-	context.subscriptions.push(vscode.commands.registerCommand(settingsCommandId, changeSettings));
+	context.subscriptions.push(vscode.commands.registerCommand(settingsCommandId, openConfig));
 	buttons.settings.text = '$(gear)';
 	buttons.settings.command = settingsCommandId;
 
@@ -169,13 +157,11 @@ export async function activate(context: vscode.ExtensionContext) {
 
 }
 
-
 class ConfigurationProvider implements vscode.DebugConfigurationProvider {
 	async resolveDebugConfiguration(folder: vscode.WorkspaceFolder | undefined, config: vscode.DebugConfiguration, token: vscode.CancellationToken): Promise<vscode.DebugConfiguration | null | undefined> {
 		if (!globalContext.globalState.get('enabled')) {
 			return new Promise(resolve => { resolve(config); });
 		}
-
 
 		if (config.__parentId) { // For some reason resolveDebugConfiguration runs twice for Node projects. __parentId is populated.
 			return new Promise(resolve => {
@@ -183,7 +169,7 @@ class ConfigurationProvider implements vscode.DebugConfigurationProvider {
 			});
 		}
 
-		if (globalContext.workspaceState.get<boolean>('telemetry', true)) {
+		if (vscode.env.isTelemetryEnabled) {
 			let lastChecked = globalContext.globalState.get('lastChecked', 0);
 			if (lastChecked < Date.now() - versionCheckInterval) {
 				checkVersion(globalContext.extension.packageJSON.version);
@@ -191,96 +177,70 @@ class ConfigurationProvider implements vscode.DebugConfigurationProvider {
 			}
 		}
 
-		const podNamespace = globalContext.workspaceState.get<string>('impersonatedPodNamespace', 'default');
-		let k8sApi = getK8sApi();
-		// Get pods from kubectl and let user select one to mirror
-		let pods: { response: any, body: V1PodList } = await k8sApi.listNamespacedPod(podNamespace);
-		let podNames = pods.body.items.map((pod) => pod.metadata!.name!);
+		let libraryPath: string;
+		if (globalContext.extensionMode === vscode.ExtensionMode.Development) {
+			libraryPath = path.join(path.dirname(globalContext.extensionPath), "target", "debug");
+		} else {
+			libraryPath = globalContext.extensionPath;
+		}
+		let [environmentVariableName, libraryName] = LIBRARIES[os.platform()];
+		config.env ||= {};
+		config.env[environmentVariableName] = path.join(libraryPath, libraryName);
 
-		await vscode.window.showQuickPick(podNames, { placeHolder: 'Select pod to mirror' }).then(async podName => {
-			return new Promise(resolve => {
-				console.log(config);
-				// Get pods from kubectl and let user select one to mirror
-				if (k8sApi === null) {
-					return;
-				}
-
-				let libraryPath;
-				if (globalContext.extensionMode === vscode.ExtensionMode.Development) {
-					libraryPath = path.join(path.dirname(globalContext.extensionPath), "target", "debug");
-				} else {
-					libraryPath = globalContext.extensionPath;
-				}
-				let [environmentVariableName, libraryName] = LIBRARIES[os.platform()];
-				globalContext.workspaceState.update('impersonatedPodName', podName);
-				config.env = {
-					...config.env, ...{
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_IMPERSONATED_TARGET': 'pod/' + podName,
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_TARGET_NAMESPACE': podNamespace,
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_AGENT_NAMESPACE': globalContext.workspaceState.get('agentNamespace', 'default'),
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_FILE_OPS': globalContext.workspaceState.get('fileOps', true).toString(),
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_FILE_RO_OPS': 'false', // TODO: Add this to settings
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_ACCEPT_INVALID_CERTIFICATES': globalContext.workspaceState.get('invalidCertificates', false).toString(),
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_AGENT_TCP_STEAL_TRAFFIC': globalContext.workspaceState.get('trafficStealing', false).toString(),
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_REMOTE_DNS': globalContext.workspaceState.get('remoteDNS', true).toString(),
-						// eslint-disable-next-line @typescript-eslint/naming-convention
-						'MIRRORD_TCP_OUTGOING': globalContext.workspaceState.get('outgoingTraffic', true).toString(),
+		if (!await isTargetInFile()) { // If target wasn't specified in the config file, let user choose pod from dropdown
+			let podNamespace: string;
+			try {
+				podNamespace = await parseNamespace();
+			} catch (e: any) {
+				vscode.window.showErrorMessage('Failed to parse mirrord config file', e.message);
+				return;
+			}
+			let k8sApi = getK8sApi();
+			// Get pods from kubectl and let user select one to mirror
+			let pods: { response: any, body: V1PodList } = await k8sApi.listNamespacedPod(podNamespace);
+			let podNames = pods.body.items.map((pod) => pod.metadata!.name!);
+			let impersonatedPodName = '';
+			await vscode.window.showQuickPick(podNames, { placeHolder: 'Select pod to mirror' }).then(async podName => {
+				return new Promise(async resolve => {
+					if (!k8sApi || !podName) {
+						return;
 					}
-				};
-
-				// mirrord doesn't support specifying both the include and exclude env vars (even if they're empty)
-				let include, exclude;
-				if (include = globalContext.workspaceState.get('includeEnvironmentVariables')) {
-					config.env['MIRRORD_OVERRIDE_ENV_VARS_INCLUDE'] = include;
-				}
-				if (exclude = globalContext.workspaceState.get('excludeEnvironmentVariables')) {
-					config.env['MIRRORD_OVERRIDE_ENV_VARS_EXCLUDE'] = exclude;
-				}
-
-				config.env[environmentVariableName] = path.join(libraryPath, libraryName);
-
-				if (config.type === "go") {
-					config.env["MIRRORD_SKIP_PROCESSES"] = "dlv;debugserver;compile;go;asm;cgo;link";
-					// use our custom delve to fix being loaded into debugserver
-					if (os.platform() === "darwin") {
-						config.dlvToolPath = path.join(globalContext.extensionPath, "dlv");
-					}
-				}
-
-				return resolve(config);
-			});
-		});
-
-		// let user select container name if there are multiple containers in the pod
-		const podName = globalContext.workspaceState.get('impersonatedPodName');
-		const pod = pods.body.items.find(p => p.metadata!.name === podName!);
-		const containerNames = pod!.spec!.containers.map(c => c.name!);
-		if (containerNames.length > 1) {
-			return await vscode.window.showQuickPick(containerNames, { placeHolder: 'Select containerName' }).then(async containerName => {
-				return new Promise(resolve => {
-					globalContext.workspaceState.update('impersonatedContainerName', containerName);
-					config.env = {
-						...config.env, ...{
-							// eslint-disable-next-line @typescript-eslint/naming-convention
-							'MIRRORD_IMPERSONATED_CONTAINER_NAME': containerName,
-						}
-					};
+					impersonatedPodName = podName;
+					config.env['MIRRORD_IMPERSONATED_TARGET'] = 'pod/' + impersonatedPodName;
+					config.env['MIRRORD_TARGET_NAMESPACE'] = podNamespace; // This is unnecessary 99% of the time, but it ensures the namespace is the one the pod was selected from
 					return resolve(config);
 				});
 			});
-		} else {
-			return new Promise(resolve => {
-				return resolve(config);
-			});
+
+			// let user select container name if there are multiple containers in the pod		
+			const pod = pods.body.items.find(p => p.metadata!.name === impersonatedPodName!);
+			const containerNames = pod!.spec!.containers.map(c => c.name!);
+			if (containerNames.length > 1) {
+				await vscode.window.showQuickPick(containerNames, { placeHolder: 'Select container' }).then(async containerName => {
+					return new Promise(resolve => {
+						config.env['MIRRORD_IMPERSONATED_CONTAINER_NAME'] = containerName;
+						return resolve(config);
+					});
+				});
+			}
 		}
+
+		let filePath = await configFilePath();
+		if (filePath) {
+			config.env['MIRRORD_CONFIG_FILE'] = filePath;
+		}
+
+		if (config.type === "go") {
+			config.env["MIRRORD_SKIP_PROCESSES"] = "dlv;debugserver;compile;go;asm;cgo;link";
+			// use our custom delve to fix being loaded into debugserver
+			if (os.platform() === "darwin") {
+				config.dlvToolPath = path.join(globalContext.extensionPath, "dlv");
+			}
+		}
+
+		return new Promise(resolve => {
+			return resolve(config);
+		});
 	}
 }